// web-ANGULAR/dfps-web-impact/src/app/modules/mytasks/neice-transmittals/enum/neice-transmittal-type.ts
export enum IncomingOutgoingTransmittals {
    Incoming_Transmittals = 'I',
    Outgoing_Transmittals = 'O',
    Outgoing_Staging_Transmittals = 'OS',
}

export enum TransmittalTypeEnum {
    HOME_STUDY_REQUEST = '150',
    HOME_STUDY_UPDATE = '160',
    HOME_STUDY_RESPONSE = '170',
    STATUS_REPORT_REQUEST = '180',
    STATUS_REPORT_RESPONSE = '190',
    PROGRESS_REPORT_REQUEST = '200',
    PROGRESS_REPORT_RESPONSE = '210',
    CONCURRENCE_REQUEST = '220',
    CONCURRENCE_RESPONSE = '230',
    PLACEMENT_DECISION = '240',
    ADDITIONALINFORMATION = '250',
}